Restrict threading for porous flow adaptivity tri test
Per https://civet.inl.gov/job/604046/ we seem to occasionally get some
crashes with this test with threading. I can't reproduce this on my
machine but that doesn't mean the bug isn't there. It could be this only
happens on a very heavily loaded slow machine (which was the case for
the CIVET crash). Maybe @cpgr can run this through helgrind or something
if he ever has time